Move training_op_helpers to core/framework.
