+++ tensorflow/core/framework/training_op_helpers.h
