+++ tensorflow/core/kernels/BUILD
@@ -161,7 +161,6 @@ tf_kernel_library(
         ":dense_update_functor",
         ":inplace_ops",
         ":ops_util",
-        ":training_op_helpers",
         ":variable_ops",
         "//tensorflow/core:framework",
         "//tensorflow/core:lib",
@@ -668,19 +667,6 @@ cc_library(
     deps = [
         ":queue_base",
         "//tensorflow/core:framework",
-    ],
-)
-
-cc_library(
-    name = "training_op_helpers",
-    srcs = ["training_op_helpers.cc"],
-    hdrs = ["training_op_helpers.h"],
-    visibility = [":friends"],
-    deps = [
-        ":dense_update_functor",
-        ":variable_ops",
-        "//tensorflow/core:framework",
-        "//tensorflow/core:lib",
     ],
 )
 
@@ -2625,7 +2611,6 @@ tf_kernel_library(
         ":gather_functor",
         ":gather_nd_op",
         ":scatter_functor",
-        ":training_op_helpers",
         ":variable_ops",
         "//tensorflow/core:core_cpu_lib",
         "//tensorflow/core:framework",
@@ -4744,7 +4729,6 @@ tf_kernel_library(
         ":resource_variable_ops",
         ":scatter_functor",
         ":state",
-        ":training_op_helpers",
         ":variable_ops",
         "//tensorflow/core:core_cpu",
         "//tensorflow/core:core_cpu_lib",
@@ -5171,7 +5155,6 @@ tf_kernel_library(
     ],
     deps = STATE_DEPS + [
         ":dense_update_functor",
-        ":training_op_helpers",
         ":variable_ops",
         ":inplace_ops",
     ],
@@ -5459,7 +5442,6 @@ tf_kernel_library(
     prefix = "training_ops",
     deps = [
         ":bounds_check",
-        ":training_op_helpers",
         ":variable_ops",
         "//tensorflow/core:framework",
         "//tensorflow/core:lib",
@@ -5554,7 +5536,6 @@ tf_kernel_library(
         ":random_ops",
         ":resource_variable_ops",
         ":stateful_random_ops",
-        ":training_op_helpers",
         "//tensorflow/core:framework",
         "//tensorflow/core:framework_internal",
         "//tensorflow/core:lib",
@@ -6080,7 +6061,6 @@ filegroup(
         "tile_ops_cpu_impl.h",
         "tile_ops_impl.h",
         "topk_op.h",
-        "training_op_helpers.h",
         "training_ops.h",
         "transpose_functor.h",
         "transpose_op.h",
@@ -6283,7 +6263,6 @@ filegroup(
         "tile_ops_cpu_impl_6.cc",
         "tile_ops_cpu_impl_7.cc",
         "topk_op.cc",
-        "training_op_helpers.cc",
         "training_ops.cc",
         "transpose_functor_cpu.cc",
         "transpose_op.cc",
